+++ shuup/admin/modules/media/static_src/media/browser/FileUpload.js
@@ -64,8 +64,12 @@ function handleFileXhrComplete(xhr, file, error) {
     var messageText = null;
     try {
         const responseJson = JSON.parse(xhr.responseText);
-        if (responseJson && responseJson.message) {
-            messageText = responseJson.message;
+        if (responseJson) {
+            if (responseJson.message) {
+                messageText = responseJson.message;
+            } else if (responseJson.error) {
+                messageText = responseJson.error;
+            }
         }
     } catch (e) {
         // invalid JSON? pffff.